refactor: renamed the home index archive to simply "Home" for better clarity
